requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Database
sqlalchemy==2.0.23
databases==0.8.0
aiosqlite==0.19.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS & Middleware
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Utilities
numpy==1.26.2
